+++ scripts/macros/spells/summonLesserDemons.js
@@ -36,7 +36,7 @@ export async function summonLesserDemons({speaker, actor, token, character, item
         return;
     }
     let userId = game.settings.get('chris-premades', 'LastGM');
-    if (game.settings.get('chris-premades', 'Player Choses Conjures')) userId = game.userId;
+    if (game.settings.get('chris-premades', 'Player Chooses Conjures')) userId = game.userId;
     if (!userId) return;
     let sourceActors = await chris.remoteDocumentsDialog(userId, 'Select Summons (Max ' + (summonsMultiplier * 2 / cr) + ')', actors);
     if (!sourceActors) return;